fixed colour and customer name
